terminal: Bitmap font can use uint8_t
The baked Deja Vu Sans Mono is 8 pixels wide, so we don't need
to use uint16_t to hold rows of pixel data. Also, splash-log
doesn't render text anymore (it defers to the kernel console),
so we don't need to include the font there any more.

+++ apps/terminal-font.h
@@ -48,7 +48,7 @@
 #define LARGE_FONT_CELL_WIDTH 8
 #define LARGE_FONT_CELL_HEIGHT 17
 #define LARGE_FONT_MASK 7
-uint16_t large_font[][17] = {
+uint8_t large_font[][17] = {
 	/* 0 32 */
 	{
 		/*          */
